bot/requirements.txt
vkbottle==4.3.8
Pillow==9.1.1
bs4==0.0.1
gTTS==2.2.4
random-dog
loguru==0.6.0
mc.py==4.0.0
Wand==0.6.7
orpl==22313
wikipedia==1.4.0
psutil==5.9.1
py-cpuinfo==8.0.0